Add hero section and menu to home page
