+++ react/src/App.tsx
@@ -3,12 +3,14 @@ import "./output.css";
 import Footer from "./Components/Footer";
 import NavBar from "./Components/NavBar";
 import { Outlet } from "react-router-dom";
+import { candles } from "./types/candles";
 
 function App() {
+  const [candlesArray, setCandlesArray] = React.useState(candles);
   return (
-    <div className="flex h-0 min-h-screen flex-col justify-between overflow-x-hidden">
+    <div className="scrollbar-none flex h-0 min-h-screen flex-col justify-between overflow-x-hidden">
       <NavBar />
-      <Outlet />
+      <Outlet context={[candlesArray, setCandlesArray]} />
       <Footer />
     </div>
   );
