+++ react/src/Components/NavBar/DesktopMenu/index.tsx
@@ -4,7 +4,7 @@ import DesktopCandlesBySizeMenu from "./DesktopCandlesBySizeMenu";
 
 const DesktopMenu = () => {
   return (
-    <nav className="top-full absolute hidden w-full flex-row-reverse items-center justify-evenly bg-white/50 py-[51px] backdrop-blur-sm xl:flex">
+    <nav className="absolute top-full hidden w-full flex-row-reverse items-center justify-evenly bg-white/50 py-[51px] backdrop-blur-sm xl:flex">
       <DesktopNavLink
         to="/"
         text="ראשי"
