+++ react/src/Components/NavBar/SideNavBar/MenuNavLink.tsx
@@ -17,7 +17,7 @@ const MenuNavLink = ({
       onClick={onClick}
       end
       className={({ isActive }) =>
-        `w-full lg:text-[32px] px-7 py-5 lg:w-7/12 lg:px-3 lg:text-center lg:font-normal lg:hover:bg-primary-pink ${isActive ? "bg-[#B0C4B1]/30 lg:bg-white/0" : "bg-white lg:bg-white/0"} ${className}`
+        `w-full px-7 py-5 lg:w-7/12 lg:px-3 lg:text-center lg:text-[32px] lg:font-normal lg:hover:bg-primary-pink ${isActive ? "bg-[#B0C4B1]/30 lg:bg-white/0" : "bg-white lg:bg-white/0"} ${className}`
       }
       to={to}
     >
