+++ react/src/Components/NavBar/SideNavBar/MenuNavOptions.tsx
@@ -16,14 +16,14 @@ const MenuNavOptions = ({
       className="flex w-full flex-col divide-y overflow-hidden lg:w-7/12"
     >
       <button
-        className="z-40 lg:justify-evenly flex w-full hover:bg-primary-pink items-center lg:gap-1 lg:font-normal gap-4 bg-white lg:px-3 px-7 py-5 text-start lg:text-[32px]"
+        className="z-40 flex w-full items-center gap-4 bg-white px-7 py-5 text-start hover:bg-primary-pink lg:justify-evenly lg:gap-1 lg:px-3 lg:text-[32px] lg:font-normal"
         onClick={toggleSizeMenu}
       >
         <p>נרות לפי גודל</p>
-        <DownArrow  sizeMenuIsOpen={sizeMenuIsOpen} />
+        <DownArrow sizeMenuIsOpen={sizeMenuIsOpen} />
       </button>
       <div
-        className={`z-10 flex w-full lg:items-center flex-col text-center text-xl font-semibold transition-all duration-300 ${sizeMenuIsOpen ? "" : "hidden"}`}
+        className={`z-10 flex w-full flex-col text-center text-xl font-semibold transition-all duration-300 lg:items-center ${sizeMenuIsOpen ? "" : "hidden"}`}
       >
         {children}
       </div>
