+++ react/src/views/Candles/Designed/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 const Designed = () => {
-  return <div className="w-full bg-red-600 h-screen">Designed</div>;
+  return <div className="h-screen w-full bg-red-600">Designed</div>;
 };
 
 export default Designed;
