+++ react/src/views/Candles/sized/[size]/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link, useOutletContext, useParams } from "react-router-dom";
+import { useOutletContext, useParams } from "react-router-dom";
 import { CandleType } from "../../../../types/candles";
 import CandleThumbnail from "../../../../Components/Candles/CandleThumbnail";
 
@@ -7,7 +7,8 @@ const BySize = () => {
   const [candlesArray, setCandlesArray] =
     useOutletContext<Array<CandleType>[]>();
   const { size } = useParams<{ size: string }>();
-  const hebrewSize = size === "small" ? "קטן" : size === "medium" ? "בינוני" : "גדול";
+  const hebrewSize =
+    size === "small" ? "קטן" : size === "medium" ? "בינוני" : "גדול";
   return (
     <div>
       Size: {size}
@@ -22,7 +23,5 @@ const BySize = () => {
     </div>
   );
 };
-
-
 
 export default BySize;
